Re-implement ingest flow via an actor rather than plain futures
This should be more flexible and consistent w/ other tasks
