+++ modules/admin/app/actors/harvesting/OaiPmhHarvester.scala
@@ -1,155 +1,122 @@
 package actors.harvesting
 
-import java.time.Instant
-import java.time.format.DateTimeFormatter
+import java.time.{Duration, LocalDateTime}
 
-import actors.harvesting.OaiPmhHarvester.OaiPmhHarvestJob
-import akka.actor.SupervisorStrategy.Stop
-import akka.actor.{Actor, ActorLogging, ActorRef, OneForOneStrategy, Props, SupervisorStrategy, Terminated}
-import models.{OaiPmhConfig, UserProfile}
-import services.harvesting.{HarvestEventHandle, HarvestEventService, OaiPmhClient, OaiPmhError}
+import actors.harvesting.OaiPmhHarvesterManager.OaiPmhHarvestJob
+import akka.actor.Status.Failure
+import akka.actor.{Actor, ActorLogging, ActorRef}
+import models.UserProfile
+import services.harvesting.OaiPmhClient
 import services.storage.FileStorage
-import utils.WebsocketConstants
 
 import scala.concurrent.ExecutionContext
 
 
 object OaiPmhHarvester {
 
-  /**
-    * A description of an OAI-PMH harvest task.
-    *
-    * @param config     the endpoint configuration
-    * @param from       the starting date and time
-    * @param classifier the storage classifier on which to save files
-    * @param prefix     the path prefix on which to save files, after
-    *                   which the item identifier will be appended
-    */
-  case class OaiPmhHarvestData(
-    config: OaiPmhConfig,
-    classifier: String,
-    prefix: String,
-    from: Option[Instant] = None,
-  )
+  // Possible states for resumption tokens:
+  sealed trait ResumptionState
+  case object Initial extends ResumptionState
+  case class Next(token: String) extends ResumptionState
+  case object Empty extends ResumptionState
+  object ResumptionState {
+    def apply(opt: Option[String]): ResumptionState = opt match {
+      case Some(t) => Next(t)
+      case _ => Empty
+    }
+  }
 
-  /**
-    * A single harvest job with a unique ID.
-    */
-  case class OaiPmhHarvestJob(repoId: String, datasetId: String, jobId: String, data: OaiPmhHarvestData)
+  // Internal message we send ourselves
+  private case class Fetch(ids: List[String], next: ResumptionState, count: Int) extends Action
 
+  // Other messages we can handle
+  sealed trait Action
+  case object Starting extends Action
+  case class Completed(total: Int, secs: Long) extends Action
+  case class Error(e: Throwable) extends Action
+  case class Resuming(token: String) extends Action
+  case class DoneFile(id: String) extends Action
+  case class Cancelled(total: Int, secs: Long) extends Action
+  case object Cancel extends Action
 }
 
-case class OaiPmhHarvester(job: OaiPmhHarvestJob, client: OaiPmhClient, storage: FileStorage, eventLog: HarvestEventService)(
-  implicit userOpt: Option[UserProfile], ec: ExecutionContext) extends Actor with ActorLogging {
-
-  import OaiPmhHarvestRunner._
+case class OaiPmhHarvester (job: OaiPmhHarvestJob, client: OaiPmhClient, storage: FileStorage)(
+    implicit userOpt: Option[UserProfile], ec: ExecutionContext) extends Actor with ActorLogging {
+  import OaiPmhHarvester._
   import akka.pattern.pipe
 
-  override def supervisorStrategy: SupervisorStrategy = OneForOneStrategy() {
-    case e =>
-      self ! Error(e)
-      Stop
-  }
-
-  // Ready state: we've received a job but won't actually start
-  // until there is a channel to talk through
   override def receive: Receive = {
-    case chan: ActorRef =>
-      log.debug("Received initial subscriber, starting...")
-      val runner = context.actorOf(Props(OaiPmhHarvestRunner(job, client, storage)))
-      context.become(running(runner, Set(chan), Option.empty))
-      runner ! Initial
+    // Start the initial harvest
+    case Initial =>
+      val msgTo = sender()
+      context.become(running(msgTo, 0, LocalDateTime.now()))
+      msgTo ! Starting
+      client.listIdentifiers(job.data.config, from = job.data.from)
+        .map { case (idents, next) => Fetch(nonDeleted(idents), ResumptionState(next), 0)}
+        .pipeTo(self)
   }
 
-  /**
-    * Running state.
-    *
-    * @param runner the harvest runner actor
-    * @param subs   a set of subscribers to message w/ updates
-    * @param handle a handle through which the job logging can be concluded
-    */
-  def running(runner: ActorRef, subs: Set[ActorRef], handle: Option[HarvestEventHandle]): Receive = {
 
-    // Add a new message subscriber
-    case chan: ActorRef =>
-      log.debug(s"Added new message subscriber, ${subs.size}")
-      context.watch(chan)
-      context.become(running(runner, subs + chan, handle))
+  // The harvest is running
+  def running(msgTo: ActorRef, done: Int, start: LocalDateTime): Receive = {
 
-    case Terminated(actor) if actor == runner =>
+    // Harvest a new batch via a resumptionToken
+    case Next(token) =>
+      msgTo ! Resuming(token)
+      client.listIdentifiers(job.data.config, resume = Some(token))
+        .map { case (idents, next) => Fetch(nonDeleted(idents), ResumptionState(next), done)}
+        .pipeTo(self)
+
+    // Harvest an individual item
+    case Fetch(id :: rest, next, count) =>
+      log.debug(s"Calling become with new total: $count")
+      context.become(running(msgTo, count, start))
+      val byteSrc = client.getRecord(job.data.config, id)
+      storage.putBytes(
+        job.data.classifier,
+        fileName(job.data.prefix, id),
+        byteSrc,
+        Some("text/xml"),
+        meta = Map(
+          "source" -> "oaipmh",
+          "oaipmh-endpoint" -> job.data.config.url,
+          "oaipmh-set" -> job.data.config.set.getOrElse(""),
+          "oaipmh-job-id" -> job.jobId
+        )
+      ).map { _ =>
+        msgTo ! DoneFile(id)
+        Fetch(rest, next, count + 1)
+      }.pipeTo(self)
+
+    // Finished a batch, start a new one
+    case Fetch(Nil, next, count) =>
+      log.debug(s"Calling become with new total: $count")
+      context.become(running(msgTo, count, start))
+      self ! next
+
+    // Finish harvesting
+    case Empty =>
+      context.stop(self)
+      msgTo ! Completed(done, time(start))
+
+    // Cancel harvest
+    case Cancel =>
+      msgTo ! Cancelled(done, time(start))
       context.stop(self)
 
-    // Remove terminated subscribers
-    case Terminated(chan) =>
-      log.debug(s"Removing subscriber: $chan")
-      context.unwatch(chan)
-      context.become(running(runner, subs - chan, handle))
-
-    // Confirmation the runner has started
-    case Starting =>
-      msg(s"Starting harvest with job id: ${job.jobId}", subs)
-      job.data.from.fold(msg("Harvesting from earliest date", subs)) { from =>
-        msg(s"Harvesting from ${DateTimeFormatter.ISO_INSTANT.format(from)}", subs)
-      }
-
-    // Cancel harvest.. here we tell the runner to exit
-    // and shut down on its termination signal...
-    case Cancel => runner ! Cancel
-
-    // The runner is continuing to harvest via a resumption token
-    case Resuming(token) => msg(s"Resuming with $token", subs)
-
-    // A file has been harvested
-    case DoneFile(id) =>
-      msg(id, subs)
-      if (handle.isEmpty) {
-        eventLog.save(job.repoId, job.datasetId, job.jobId).pipeTo(self)
-      }
-
-      // We've received a log handle which we can use to say how
-      // this job finished: via error, cancellation, or otherwise
-    case handle: HarvestEventHandle =>
-      context.become(running(runner, subs, Some(handle)))
-
-    // Received confirmation that the runner has shut down
-    case Cancelled(count, secs) =>
-      msg(s"${WebsocketConstants.ERR_MESSAGE}: cancelled after $count file(s) in $secs seconds", subs)
-      handle.foreach(_.cancel())
-      context.stop(self)
-
-    // The runner has completed, so we log the
-    // event and shut down too
-    case Completed(count, secs) =>
-      msg(s"${WebsocketConstants.DONE_MESSAGE}: " +
-        s"harvested $count file(s) in $secs seconds", subs)
-      handle.foreach(_.close())
-      context.stop(self)
-
-    // Error case where the `set` or `from` parameters mean that
-    // no records are returned
-    case OaiPmhError("noRecordsMatch", _) =>
-      msg(s"${WebsocketConstants.DONE_MESSAGE}: nothing to harvest", subs)
-      context.stop(self)
-
-    // Error case where we get some other problem...
-    case e: OaiPmhError =>
-      msg(s"${WebsocketConstants.ERR_MESSAGE}: ${e.code}", subs)
-      handle.map(_.error(e))
-      context.stop(self)
-
-    // The runner has thrown an unexpected error. Log the event
-    // and shut down
-    case Error(e) =>
-      msg(s"${WebsocketConstants.ERR_MESSAGE}: harvesting error: ${e.getMessage}", subs)
-      handle.foreach(_.error(e))
+    case Failure(e) =>
+      msgTo ! e
       context.stop(self)
 
     case m =>
-      log.error(s"Unexpected message: $m")
+      log.error(s"Unexpected message: $m: ${m.getClass}")
   }
 
-  private def msg(s: String, subs: Set[ActorRef]): Unit = {
-    log.info(s + s" (subscribers: ${subs.size})")
-    subs.foreach(_ ! s)
-  }
+  private def time(from: LocalDateTime): Long =
+    Duration.between(from, LocalDateTime.now()).toMillis / 1000
+
+  private def fileName(prefix: String, id: String): String = prefix + id + ".xml"
+
+  private def nonDeleted(idents: Seq[(String, Boolean)]): List[String] = idents
+    .filterNot(_._2).map(_._1).toList
 }
