+++ test/resources/data-transformation-fixtures.sql
@@ -1,9 +1,6 @@
-INSERT INTO import_dataset VALUES (
-    'r1', 'default', 'Default', 'upload', '2020-10-09 10:00:00', 'test'
-), (
-    'r2', 'default', 'Default', 'oaipmh', '2020-10-09 10:00:00', 'test'
-);
+INSERT INTO import_dataset (repo_id, id, name, type, comments)
+VALUES ('r1', 'default', 'Default', 'upload', 'test'),
+       ('r2', 'default', 'Default', 'oaipmh', 'test');
 
-INSERT INTO data_transformation VALUES (
-  'FG8jdRd43j', 'test', 'r1','xquery','test','2020-06-12 10:00:00','test'
-);
+INSERT INTO data_transformation (id, name, repo_id, type, map, comments)
+VALUES ('FG8jdRd43j', 'test', 'r1', 'xquery', 'test', 'test');
