+++ test/resources/oaipmh-config-fixtures.sql
@@ -1,5 +1,5 @@
-INSERT INTO import_dataset
-VALUES ('r1', 'default', 'Default', 'oaipmh', '2020-10-09 10:00:00', 'test');
+INSERT INTO import_dataset (repo_id, id, name, type, comments)
+VALUES ('r1', 'default', 'Default', 'oaipmh', 'test');
 
 INSERT INTO oaipmh_config
 VALUES ('r1', 'default', 'http://example.com', 'ead', NULL, '2020-06-12 10:00:00', NULL);
